requirements.txt
asgiref==3.7.2
attrs==23.2.0
black==24.1.1
click==8.1.7
colorama==0.4.6
Django==5.0.2
django-cors-headers==4.3.1
django-jazzmin==2.6.0
django-otp==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
PyJWT==2.8.0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.17.1
sqlparse==0.4.4
tzdata==2023.4
uritemplate==4.1.1
